Implement parsing of the input file
The input file provided by the user will be parsed upon running the
program. The result is a vector of ordered pairs, each representing a
point in the xy-plane. The next step is to convert this set of points
into a graph to be used by the program to compute the length of paths.