Fix edge crossover algorithm
Edge crossover is broken because it is unable to handle the situation
where the entire genotype has been scanned and no more adjacent vertices
are available. With this patch, it will correctly make a random choice
for the next vertex to add to the child in the event that not
adjacencies exist.