Implement main evolutionary loop and option config
This completes the project in the strictest sense.
Command-line arguments determine the population size and number of
fitness evaluations before termination.
Parent selection operator is based on linear probability distribution.
Survival selection is simply replace worst.
Fitness is the distance travelled across the sequence of cities
represented by the genotype PLUS the distance from the last city to the
starting city.
Obviously much of this is hard-coded to deal with this particular
problem. But there are seeds of generic design here that can be expanded
to turn this into a general-use package for genetic algorithms. And
where no such seeds exist it is clearly possible to conceptualize from
scratch by experimentation.
As a side note: the main program outputs the best fitness of every ten
generations, each on a separate line. This could be useful for plugging
the data into something like MATLAB and graphing the progress, upon
changing certain parameters.